LG-9756: Create a document capture concern for post-FSM world (#8404)
* move save_proofing_components to new DocumentCaptureConcern

* add methods to document_capture_concern

* move stored_result method from document_capture_controllers to document_capture_concern

* remove hybrid_flow_mobile? and user_id_from_token methods in doc capture controller


[skip changelog]

* remove comment

* remove comments as methods were unchanged

* rename track_document_state to track_document_issuing_state

* remove effective user from document capture concern (#8451)


* document capture controller use document_capture_session_uuid for consistency (#8467)

* use document_capture_session_uuid for consistency

* [skip changelog]

---------


---------
