+++ app/controllers/idv/document_capture_controller.rb
@@ -1,6 +1,7 @@
 module Idv
   class DocumentCaptureController < ApplicationController
     include AcuantConcern
+    include DocumentCaptureConcern
     include IdvSession
     include IdvStepConcern
     include StepIndicatorConcern
@@ -37,7 +38,7 @@ module Idv
 
     def extra_view_variables
       {
-        document_capture_session_uuid: flow_session[:document_capture_session_uuid],
+        document_capture_session_uuid: document_capture_session_uuid,
         flow_path: 'standard',
         sp_name: decorated_session.sp_name,
         failure_to_proof_url: return_to_sp_failure_to_proof_url(step: 'document_capture'),
@@ -73,94 +74,16 @@ module Idv
       }.merge(**acuant_sdk_ab_test_analytics_args)
     end
 
-    def in_person_cta_variant_testing_variables
-      bucket = AbTests::IN_PERSON_CTA.bucket(flow_session[:document_capture_session_uuid])
-      session[:in_person_cta_variant] = bucket
-      {
-        in_person_cta_variant_testing_enabled:
-        IdentityConfig.store.in_person_cta_variant_testing_enabled,
-        in_person_cta_variant_active: bucket,
-      }
-    end
-
     def handle_stored_result
       if stored_result&.success?
-        save_proofing_components
-        extract_pii_from_doc(stored_result, store_in_session: !hybrid_flow_mobile?)
+        save_proofing_components(current_user)
+        extract_pii_from_doc(current_user, stored_result, store_in_session: true)
         flash[:success] = t('doc_auth.headings.capture_complete')
         successful_response
       else
         extra = { stored_result_present: stored_result.present? }
         failure(I18n.t('doc_auth.errors.general.network_error'), extra)
       end
-    end
-
-    def stored_result
-      return @stored_result if defined?(@stored_result)
-      @stored_result = document_capture_session&.load_result
-    end
-
-    def save_proofing_components
-      return unless current_user
-
-      doc_auth_vendor = DocAuthRouter.doc_auth_vendor(
-        discriminator: document_capture_session_uuid,
-        analytics: analytics,
-      )
-
-      component_attributes = {
-        document_check: doc_auth_vendor,
-        document_type: 'state_id',
-      }
-      ProofingComponent.create_or_find_by(user: current_user).update(component_attributes)
-    end
-
-    def hybrid_flow_mobile?
-      user_id_from_token.present?
-    end
-
-    def user_id_from_token
-      flow_session[:doc_capture_user_id]
-    end
-
-    # copied from doc_auth_base_step.rb
-    # @param [DocAuth::Response,
-    #   DocumentCaptureSessionAsyncResult,
-    #   DocumentCaptureSessionResult] response
-    def extract_pii_from_doc(response, store_in_session: false)
-      pii_from_doc = response.pii_from_doc.merge(
-        uuid: effective_user.uuid,
-        phone: effective_user.phone_configurations.take&.phone,
-        uuid_prefix: ServiceProvider.find_by(issuer: sp_session[:issuer])&.app_id,
-      )
-
-      flow_session[:had_barcode_read_failure] = response.attention_with_barcode?
-      if store_in_session
-        flow_session[:pii_from_doc] ||= {}
-        flow_session[:pii_from_doc].merge!(pii_from_doc)
-        idv_session.clear_applicant!
-      end
-      track_document_state(pii_from_doc[:state])
-    end
-
-    def track_document_state(state)
-      return unless IdentityConfig.store.state_tracking_enabled && state
-      doc_auth_log = DocAuthLog.find_by(user_id: current_user.id)
-      return unless doc_auth_log
-      doc_auth_log.state = state
-      doc_auth_log.save!
-    end
-
-    def successful_response
-      FormResponse.new(success: true)
-    end
-
-    # copied from Flow::Failure module
-    def failure(message, extra = nil)
-      flow_session[:error_message] = message
-      form_response_params = { success: false, errors: { message: message } }
-      form_response_params[:extra] = extra unless extra.nil?
-      FormResponse.new(**form_response_params)
     end
   end
 end
